feat : Header 컴포넌트 UI 수정
