+++ src/components/Header/index.tsx
@@ -1,16 +1,31 @@
-import SideBar from '../SideBar';
+import { useLocation, useNavigate } from 'react-router-dom';
 import useToggle from '../../hooks/useToggle';
-import TopHeader from './components/TopHeader';
-import BottomHeader from './components/BottomHeader';
+
+import SideBar from '../SideBar';
+import Icons from './components/Icons';
+
+import { ReactComponent as LogoBanner } from '../../assets/logo font.svg';
+import NavBar from './components/NavBar';
+import HeaderMenuToggle from './components/HeaderMenuToggle';
 
 const Header = () => {
   const [menuToggle, handleMenuToggle] = useToggle(false);
+  const { pathname } = useLocation();
+  const nav = useNavigate();
 
   return (
     <>
-      <header className="flex flex-col">
-        <TopHeader />
-        <BottomHeader menuToggle={menuToggle} handleMenuToggle={handleMenuToggle} />
+      <header className="flex justify-between items-center w-full pl-[21.31rem] pr-[20.25rem] h-[6.25rem] sm:px-10">
+        <LogoBanner onClick={() => nav('/')} className="w-[8rem] cursor-pointer" />
+        <div className="flex w-[30rem] justify-between items-center sm:hidden">
+          <NavBar pathname={pathname} navHome={() => nav('/')} navPlace={() => nav('place')} />
+          <Icons navMypage={() => nav('/mypage')} handleLogout={() => alert('logout')} />
+        </div>
+        <HeaderMenuToggle
+          menuToggle={menuToggle}
+          className="lg:hidden"
+          onClick={handleMenuToggle}
+        />
       </header>
       <SideBar menuToggle={menuToggle} handleMenuToggle={handleMenuToggle} />
     </>
